Move wasm-link into a C++ file
